requirements.txt
# Core Discord bot requirements
py-cord>=2.4.0
discord.py-voice
PyNaCl>=1.5.0

# Audio processing
ffmpeg-python>=0.2.0

# YouTube and media downloading
yt-dlp>=2023.12.30

# Web scraping and parsing
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0

# Music metadata and information
musicbrainzngs>=0.7.1

# Text similarity for autoplay
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0

# Web panel and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
starlette>=0.27.0

# Authentication and sessions
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-decouple>=3.8

# File handling and async operations
aiofiles>=23.2.0

# System monitoring
psutil>=5.9.0

# Voice recognition (optional)
coqui-stt>=1.4.0

# Data export formats
PyYAML>=6.0.1
toml>=0.10.2

# Development and debugging
python-dotenv>=1.0.0